Differential equation for von Bertalanffy growth single and multi- individual models

Differential equation for von Bertalanffy growth single and multi- individual models

Usage

hmde_vb_de(y = NULL, pars = NULL)

Arguments

y

input real

pars

list of parameteters Y_max, growth_rate

Value

value of differential equation at y